+++ examples/cursor.rs
@@ -32,6 +32,11 @@ fn main() -> Result<(), impl std::error::Error> {
                 } => {
                     println!("Setting cursor to \"{:?}\"", CURSORS[cursor_idx]);
                     window.set_cursor_icon(CURSORS[cursor_idx]);
+                    #[cfg(windows_platform)]
+                    {
+                        use winit::platform::windows::WindowExtWindows;
+                        window.set_taskbar_progress(cursor_idx as f32 / CURSORS.len() as f32);
+                    }
                     if cursor_idx < CURSORS.len() - 1 {
                         cursor_idx += 1;
                     } else {